+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Enhancements
     - Default upstream database version is now `v6.0.0`
+    - Parser results can be passed directly to `UAInspector.bot?/1`
     - Upstream improvements to detect engine versions of "Blink" type browsers have been integrated
 
 ## v3.0.1 (2022-03-08)